\name{pa-package}
\alias{pa-package}
\docType{package}
\title{
  Brinson analysis for equity portfolios
}
\description{

  The pa package provides facilities for conducting performance
  attribution based on the Brinson model and the regression approach.

}
\details{
  \tabular{ll}{
    Package: \tab pa\cr
    Type: \tab Package\cr
    Version: \tab 1.2-2\cr
    Date: \tab 2012-05-14\cr
    Depends: \tab R (>= 2.10), methods, ggplot2, grid\cr
    License: \tab GPL-2\cr
    LazyLoad: \tab yes\cr
  }

}

\author{

  Yang Lu <yang.lu2014@gmail.com>
  David Kane <dave.kane@gmail.com>

  Maintainer: Yang Lu <yang.lu2014@gmail.com>
}
\keyword{ package }