\name{pa-package}
\alias{pa-package}
\docType{package}
\title{
  Brinson analysis for equity portfolios
}
\description{

  The pa package provides facilities for conducting performance
  attribution based on the Brinson model and the regression approach.

}
\author{

  Yang Lu <yang.lu2014@gmail.com>
  David Kane <dave.kane@gmail.com>

  Maintainer: Yang Lu <yang.lu2014@gmail.com>
}
\keyword{ package }